A data frame with information about the nature and sustainability from the viewpoint of respondents to the survey

A dataset containing information about Nature and sustainability from the viewpoint of users of active and motorized modes of transportation in Santiago sourced from 2016 survey.

Usage

data(Santiago_NS)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 451 rows and 10 columns.

Source

A face to face survey considering pre-census 2012

Details

@format A data frame with 451 rows and 10 variables:

ID

Unique identifier of respondent

r5A_CHANGE

In a climatic event like heavy rain or flood, do respondents change their main mode of transport?; this variable is an out of order factor with two levels, "YES", "NO"

r5B_CHANGE_MODE

If respondents answered yes (change mode of transport in a weather event), indication of the mode(s) is required; this variable is an out of order factor with eight levels, "CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"

r5C_SUST

How do respondents assess their access to the currently available sustainable modes of transport? (eg hybrid buses, electric cars, public bicycles); this variable is an ordered factor with five levels, "POOR", "FAIR", "GOOD", "VERY GOOD", "EXCELLENT"

r5D_IMP_SUST

How important is it for respondents to have access to sustainable modes of transport?; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"

r5E_PAYMENT

Would respondents be willing to spend more on transportation to gain access to more sustainable modes?; this variable is an out of order factor with two levels, "YES", "NO"

r5F_PAYMENTS

If respondents' previous response was positive, indication of how much their transportation expense would increase; this variable is an ordered factor with three levels, "5-15%", "15-30%", "30% or more"

r5GA_TREE

indication of how important it is improving the following aspects in public transport routes: Presence of trees; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"

r5GB_PARK

indication of how important it is improving the following aspects in public transport routes: Access to parks; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"

r5GC_MODE

indication of how important it is improving the following aspects in public transport routes: Access to sustainable transport modes; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"

r5GD_MODE

indication of how important it is improving the following aspects in public transport routes: Broaden supply of sustainable transport modes; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"

Examples

 data(Santiago_NS)
 r5A_CHANGE <- Santiago_NS$r5A_CHANGE